% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotcli_helpers.r
\name{add_legend_to_output}
\alias{add_legend_to_output}
\title{Add Legend to Output Matrix}
\usage{
add_legend_to_output(output, legend_info, position, top_margin, plot_height)
}
\description{
Add Legend to Output Matrix
}
\keyword{internal}
