% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.r
\name{AsciiCanvas}
\alias{AsciiCanvas}
\title{ASCII Canvas Class}
\description{
ASCII Canvas Class

ASCII Canvas Class
}
\details{
Simple canvas using basic ASCII characters.
Resolution: 1x1 (one pixel per character)
}
\section{Super class}{
\code{\link[plotcli:Canvas]{plotcli::Canvas}} -> \code{AsciiCanvas}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{point_char}}{Character used for points}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AsciiCanvas-new}{\code{AsciiCanvas$new()}}
\item \href{#method-AsciiCanvas-set_pixel}{\code{AsciiCanvas$set_pixel()}}
\item \href{#method-AsciiCanvas-clone}{\code{AsciiCanvas$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="apply_colors"><a href='../../plotcli/html/Canvas.html#method-Canvas-apply_colors'><code>plotcli::Canvas$apply_colors()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="clear"><a href='../../plotcli/html/Canvas.html#method-Canvas-clear'><code>plotcli::Canvas$clear()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="draw_circle"><a href='../../plotcli/html/Canvas.html#method-Canvas-draw_circle'><code>plotcli::Canvas$draw_circle()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="draw_hline"><a href='../../plotcli/html/Canvas.html#method-Canvas-draw_hline'><code>plotcli::Canvas$draw_hline()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="draw_line"><a href='../../plotcli/html/Canvas.html#method-Canvas-draw_line'><code>plotcli::Canvas$draw_line()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="draw_points"><a href='../../plotcli/html/Canvas.html#method-Canvas-draw_points'><code>plotcli::Canvas$draw_points()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="draw_polygon"><a href='../../plotcli/html/Canvas.html#method-Canvas-draw_polygon'><code>plotcli::Canvas$draw_polygon()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="draw_polyline"><a href='../../plotcli/html/Canvas.html#method-Canvas-draw_polyline'><code>plotcli::Canvas$draw_polyline()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="draw_rect"><a href='../../plotcli/html/Canvas.html#method-Canvas-draw_rect'><code>plotcli::Canvas$draw_rect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="draw_segment"><a href='../../plotcli/html/Canvas.html#method-Canvas-draw_segment'><code>plotcli::Canvas$draw_segment()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="draw_text"><a href='../../plotcli/html/Canvas.html#method-Canvas-draw_text'><code>plotcli::Canvas$draw_text()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="draw_vline"><a href='../../plotcli/html/Canvas.html#method-Canvas-draw_vline'><code>plotcli::Canvas$draw_vline()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="fill_area"><a href='../../plotcli/html/Canvas.html#method-Canvas-fill_area'><code>plotcli::Canvas$fill_area()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="fill_bar"><a href='../../plotcli/html/Canvas.html#method-Canvas-fill_bar'><code>plotcli::Canvas$fill_bar()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="fill_circle"><a href='../../plotcli/html/Canvas.html#method-Canvas-fill_circle'><code>plotcli::Canvas$fill_circle()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="fill_rect"><a href='../../plotcli/html/Canvas.html#method-Canvas-fill_rect'><code>plotcli::Canvas$fill_rect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="print"><a href='../../plotcli/html/Canvas.html#method-Canvas-print'><code>plotcli::Canvas$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="plotcli" data-topic="Canvas" data-id="render"><a href='../../plotcli/html/Canvas.html#method-Canvas-render'><code>plotcli::Canvas$render()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AsciiCanvas-new"></a>}}
\if{latex}{\out{\hypertarget{method-AsciiCanvas-new}{}}}
\subsection{Method \code{new()}}{
Initialize ASCII canvas
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AsciiCanvas$new(width, height, point_char = "*")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{width}}{Character width}

\item{\code{height}}{Character height}

\item{\code{point_char}}{Character to use for points (default "*")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AsciiCanvas-set_pixel"></a>}}
\if{latex}{\out{\hypertarget{method-AsciiCanvas-set_pixel}{}}}
\subsection{Method \code{set_pixel()}}{
Set a pixel
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AsciiCanvas$set_pixel(x, y, color = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{X coordinate (1-based)}

\item{\code{y}}{Y coordinate (1-based, 1 = top)}

\item{\code{color}}{Optional color name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AsciiCanvas-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AsciiCanvas-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AsciiCanvas$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
