% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{braille_dot_bit}
\alias{braille_dot_bit}
\title{Get Braille dot bit value}
\usage{
braille_dot_bit(dot_row, dot_col)
}
\arguments{
\item{dot_row}{Row within the Braille cell (0-3, top to bottom)}

\item{dot_col}{Column within the Braille cell (0-1, left to right)}
}
\value{
The bit value for that dot position
}
\description{
Returns the bit value for a specific dot position in a Braille cell.
Braille cells have a 2x4 dot matrix with the following bit values:
}
\details{
\preformatted{
       Col 0  Col 1
Row 0:  0x01   0x08
Row 1:  0x02   0x10
Row 2:  0x04   0x20
Row 3:  0x40   0x80
}
}
