% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{braille_set_dot}
\alias{braille_set_dot}
\title{Set a dot in a Braille character}
\usage{
braille_set_dot(current_char, dot_row, dot_col)
}
\arguments{
\item{current_char}{The current character (can be space or existing Braille)}

\item{dot_row}{Row within the Braille cell (0-3)}

\item{dot_col}{Column within the Braille cell (0-1)}
}
\value{
The updated Braille character
}
\description{
Set a dot in a Braille character
}
