% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{bresenham}
\alias{bresenham}
\title{Bresenham's line algorithm}
\usage{
bresenham(x0, y0, x1, y1)
}
\arguments{
\item{x0}{The x-coordinate of the starting point.}

\item{y0}{The y-coordinate of the starting point.}

\item{x1}{The x-coordinate of the ending point.}

\item{y1}{The y-coordinate of the ending point.}
}
\value{
A list of points that form a line between the two given points.
}
\description{
This function generates a list of points that form a line between two given points using Bresenham's line algorithm.
}
\examples{
bresenham(0, 0, 5, 5)
bresenham(0, 0, -5, -5)
}
