% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotcli_helpers.r
\name{build_plot_output_v2}
\alias{build_plot_output_v2}
\title{Build Plot Output with Axes and Title (v2)}
\usage{
build_plot_output_v2(
  canvas,
  scales,
  width,
  height,
  style_opts,
  left_margin,
  top_margin,
  legend_info = NULL
)
}
\arguments{
\item{canvas}{The rendered canvas}

\item{scales}{The scales object}

\item{width}{Total width}

\item{height}{Total height}

\item{style_opts}{Style options}

\item{left_margin}{Left margin size}

\item{top_margin}{Top margin size}

\item{legend_info}{Legend information from extract_legend_info}
}
\value{
Character matrix
}
\description{
Build Plot Output with Axes and Title (v2)
}
\keyword{internal}
