% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{cat_plot_matrix}
\alias{cat_plot_matrix}
\title{Print plot matrix}
\usage{
cat_plot_matrix(plot_matrix)
}
\arguments{
\item{plot_matrix}{The plot matrix to be printed.}
}
\description{
This function prints a plot matrix to the console.
}
\examples{
cat_plot_matrix(matrix(c("a", "b", "c", "d"), nrow = 2, ncol = 2))
}
