% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcli.r
\name{cbind.plotcli}
\alias{cbind.plotcli}
\title{Combine two plotcli objects horizontally}
\usage{
\method{cbind}{plotcli}(plot1, plot2)
}
\arguments{
\item{plot1}{A plotcli object to be combined.}

\item{plot2}{A plotcli object to be combined.}
}
\value{
A combined plot matrix.
}
\description{
This function combines two plotcli objects horizontally, centering them vertically.
}
