% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.r
\name{cbind_plots}
\alias{cbind_plots}
\title{Combine plot matrices horizontally}
\usage{
cbind_plots(...)
}
\arguments{
\item{...}{A list of plot matrices to be combined.}
}
\value{
A combined plot matrix.
}
\description{
This function combines multiple plot matrices horizontally, centering them vertically.
}
