% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.r
\name{create_canvas}
\alias{create_canvas}
\title{Create a canvas of the specified type}
\usage{
create_canvas(width, height, type = "braille")
}
\arguments{
\item{width}{Character width}

\item{height}{Character height}

\item{type}{Canvas type: "ascii", "braille", or "block"}
}
\value{
A Canvas object
}
\description{
Create a canvas of the specified type
}
