% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\docType{data}
\name{.geom_registry}
\alias{.geom_registry}
\title{Geom Registry Environment}
\format{
An object of class \code{environment} of length 15.
}
\usage{
.geom_registry
}
\description{
Internal environment storing registered geom handlers
}
\keyword{internal}
