% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotcli_helpers.r
\name{extract_legend_info}
\alias{extract_legend_info}
\title{Extract Legend Information from ggplot}
\usage{
extract_legend_info(built)
}
\description{
Extract Legend Information from ggplot
}
\keyword{internal}
