% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotcli_helpers.r
\name{extract_plot_labels}
\alias{extract_plot_labels}
\title{Extract Plot Labels from ggplot}
\usage{
extract_plot_labels(built, title, subtitle, caption, axis_labels)
}
\description{
Extract Plot Labels from ggplot
}
\keyword{internal}
