% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotcli_helpers.r
\name{extract_plot_style}
\alias{extract_plot_style}
\title{Extract Plot Style from ggplot Theme}
\usage{
extract_plot_style(built, border, grid, legend)
}
\description{
Extract Plot Style from ggplot Theme
}
\keyword{internal}
