% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotcli_helpers.r
\name{format_axis_label}
\alias{format_axis_label}
\title{Format Axis Label}
\usage{
format_axis_label(value)
}
\arguments{
\item{value}{Numeric value}
}
\value{
Formatted string
}
\description{
Format Axis Label
}
\keyword{internal}
