% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{format_four_chars}
\alias{format_four_chars}
\title{Format number to four characters}
\usage{
format_four_chars(num)
}
\arguments{
\item{num}{The number to be formatted.}
}
\value{
A string representation of the number with exactly four characters.
}
\description{
This function formats a number to a string of exactly four characters.
}
\examples{
format_four_chars(123)
format_four_chars(-12.34)
}
