% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{geom_area_handler}
\alias{geom_area_handler}
\title{GeomArea Handler}
\usage{
geom_area_handler(data, canvas, scales, params, style_opts = NULL)
}
\description{
Renders filled areas
}
\keyword{internal}
