% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{geom_bar_handler}
\alias{geom_bar_handler}
\title{GeomBar/GeomCol Handler}
\usage{
geom_bar_handler(data, canvas, scales, params, style_opts = NULL)
}
\description{
Renders bar charts
}
\keyword{internal}
