% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{geom_boxplot_handler}
\alias{geom_boxplot_handler}
\title{GeomBoxplot Handler}
\usage{
geom_boxplot_handler(data, canvas, scales, params)
}
\description{
Renders boxplots with whiskers, box, median line, and outliers
}
\keyword{internal}
