% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{geom_density_handler}
\alias{geom_density_handler}
\title{GeomDensity Handler}
\usage{
geom_density_handler(data, canvas, scales, params, style_opts = NULL)
}
\description{
Renders density curves
}
\keyword{internal}
