% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{geom_histogram_handler}
\alias{geom_histogram_handler}
\title{GeomHistogram Handler}
\usage{
geom_histogram_handler(data, canvas, scales, params, style_opts = NULL)
}
\description{
Renders histograms (same as bar)
}
\keyword{internal}
