% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{geom_hline_handler}
\alias{geom_hline_handler}
\title{GeomHline Handler}
\usage{
geom_hline_handler(data, canvas, scales, params, style_opts = NULL)
}
\description{
Renders horizontal lines
}
\keyword{internal}
