% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{geom_line_handler}
\alias{geom_line_handler}
\title{GeomLine Handler}
\usage{
geom_line_handler(data, canvas, scales, params, style_opts = NULL)
}
\description{
Renders connected lines
}
\keyword{internal}
