% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{geom_path_handler}
\alias{geom_path_handler}
\title{GeomPath Handler}
\usage{
geom_path_handler(data, canvas, scales, params, style_opts = NULL)
}
\description{
Renders connected paths (order by data, not x)
}
\keyword{internal}
