% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{geom_point_handler}
\alias{geom_point_handler}
\title{GeomPoint Handler}
\usage{
geom_point_handler(data, canvas, scales, params, style_opts = NULL)
}
\description{
Renders points as individual pixels or small shapes
}
\keyword{internal}
