% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{geom_rect_handler}
\alias{geom_rect_handler}
\title{GeomRect Handler}
\usage{
geom_rect_handler(data, canvas, scales, params, style_opts = NULL)
}
\description{
Renders rectangles
}
\keyword{internal}
