% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{geom_segment_handler}
\alias{geom_segment_handler}
\title{GeomSegment Handler}
\usage{
geom_segment_handler(data, canvas, scales, params, style_opts = NULL)
}
\description{
Renders line segments
}
\keyword{internal}
