% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{geom_smooth_handler}
\alias{geom_smooth_handler}
\title{GeomSmooth Handler}
\usage{
geom_smooth_handler(data, canvas, scales, params, style_opts = NULL)
}
\description{
Renders smoothed lines (just draws the line, ignores confidence interval)
}
\keyword{internal}
