% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{geom_text_handler}
\alias{geom_text_handler}
\title{GeomText Handler}
\usage{
geom_text_handler(data, canvas, scales, params, style_opts = NULL)
}
\description{
Renders text labels
}
\keyword{internal}
