% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{geom_vline_handler}
\alias{geom_vline_handler}
\title{GeomVline Handler}
\usage{
geom_vline_handler(data, canvas, scales, params, style_opts = NULL)
}
\description{
Renders vertical lines
}
\keyword{internal}
