% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{get_color_hue}
\alias{get_color_hue}
\title{Get the hue of a color (0-360 degrees)}
\usage{
get_color_hue(color)
}
\arguments{
\item{color}{A color value}
}
\value{
Hue in degrees (0-360) or NA for grayscale
}
\description{
Get the hue of a color (0-360 degrees)
}
\keyword{internal}
