% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotcli.r
\name{get_data_subset}
\alias{get_data_subset}
\title{Get data subset for a specific geom}
\usage{
get_data_subset(geom_name, data, aes, p_build)
}
\arguments{
\item{geom_name}{The name of the geom for which the data subset is needed.}

\item{data}{The data to be subsetted.}

\item{aes}{The aesthetic mappings for the geom.}

\item{p_build}{The ggplot build object.}
}
\value{
A list containing the data subset for the specified geom.
}
\description{
This function returns a subset of the data for a specific geom.
}
