% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotcli_helpers.r
\name{get_facet_info}
\alias{get_facet_info}
\title{Get Facet Information from Layout}
\usage{
get_facet_info(layout)
}
\arguments{
\item{layout}{The layout object from ggplot_build}
}
\value{
List with facet info
}
\description{
Get Facet Information from Layout
}
\keyword{internal}
