% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{get_term_colors}
\alias{get_term_colors}
\title{Get terminal colors}
\usage{
get_term_colors(n = NULL)
}
\arguments{
\item{n}{The number of colors to return.}
}
\value{
A vector of terminal colors.
}
\description{
This function returns a vector of terminal colors.
}
\examples{
get_term_colors(5)
get_term_colors(10)
}
