% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotcli.r
\name{ggplotcli}
\alias{ggplotcli}
\title{ggplotcli - Render ggplot objects in the terminal}
\usage{
ggplotcli(ggplot_obj, plot_width = 80, plot_height = 40, braille = TRUE)
}
\arguments{
\item{ggplot_obj}{A ggplot object to be rendered in the terminal.}

\item{plot_width}{Width of the terminal plot in characters (default: 80).}

\item{plot_height}{Height of the terminal plot in characters (default: 40).}

\item{braille}{Use Braille characters for higher resolution (default: TRUE).}
}
\value{
A TerminalPlot object.
}
\description{
This function takes a ggplot object and renders it in the terminal using ASCII or Braille characters.
}
