% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{init_color_mapping}
\alias{init_color_mapping}
\title{Initialize color mapping for a set of ggplot colors}
\usage{
init_color_mapping(ggplot_colors)
}
\arguments{
\item{ggplot_colors}{Vector of unique colors from ggplot}
}
\description{
This function takes all unique colors from a ggplot and assigns terminal
colors to minimize repetition while respecting hue similarity.
}
