% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{is_braille}
\alias{is_braille}
\title{Check if a character is a Braille character}
\usage{
is_braille(char)
}
\arguments{
\item{char}{The character to be checked.}
}
\value{
A boolean value indicating whether the character is a Braille character or not.
}
\description{
This function checks if a given character is a Braille character.
}
\examples{
is_braille("A")
}
