% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{is_geom_registered}
\alias{is_geom_registered}
\title{Check if a Geom is Registered}
\usage{
is_geom_registered(geom_name)
}
\arguments{
\item{geom_name}{Name of the geom}
}
\value{
Logical
}
\description{
Check if a Geom is Registered
}
