% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{list_registered_geoms}
\alias{list_registered_geoms}
\title{List Registered Geoms}
\usage{
list_registered_geoms()
}
\value{
Character vector of registered geom names
}
\description{
List Registered Geoms
}
