% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{make_colored}
\alias{make_colored}
\title{Make colored text}
\usage{
make_colored(x, color = NULL)
}
\arguments{
\item{x}{The text string to be colored.}

\item{color}{The color to be applied to the text. If NULL, the color codes will be removed.}
}
\value{
A colored text string or a text string with color codes removed.
}
\description{
This function applies a specified color to a given text string.
}
\examples{
make_colored("Hello, world!", "blue")
make_colored("Hello, world!", NULL)
}
