% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{make_unique_names}
\alias{make_unique_names}
\title{Make unique names}
\usage{
make_unique_names(names)
}
\arguments{
\item{names}{A character vector of names.}
}
\value{
A character vector of unique names.
}
\description{
This function takes a vector of names and ensures that each name is unique by appending a number if necessary.
}
\examples{
make_unique_names(c("apple", "apple", "banana", "apple"))
}
