% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{normalize_data}
\alias{normalize_data}
\title{Normalize data}
\usage{
normalize_data(data, data_min, data_max, plot_range)
}
\arguments{
\item{data}{The data to be normalized.}

\item{data_min}{The minimum value of the data.}

\item{data_max}{The maximum value of the data.}

\item{plot_range}{The range to normalize the data to.}
}
\value{
The normalized data.
}
\description{
This function normalizes the given data to a specified plot range.
}
\examples{
normalize_data(c(1, 2, 3, 4, 5), 1, 5, 10)
normalize_data(c(10, 20, 30, 40, 50), 10, 50, 100)
}
