% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcli_wrappers.r
\name{pclib}
\alias{pclib}
\title{Short version of plotcli_bar}
\usage{
pclib(
  y,
  x = NULL,
  plot_width = 80,
  plot_height = 40,
  x_label = "x",
  y_label = "y",
  color = NULL,
  braille = TRUE,
  name = "barplot",
  ...
)
}
\description{
Short version of plotcli_bar function.
}
