% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcli_wrappers.r
\name{pclibx}
\alias{pclibx}
\title{Short version of plotcli_box}
\usage{
pclibx(
  y,
  plot_width = getOption("plotcli.plot_width", 80),
  plot_height = getOption("plotcli.plot_height", 40),
  x_label = "x",
  y_label = "y",
  color = NULL,
  braille = getOption("plotcli.braille", TRUE),
  name = "boxplot",
  ...
)
}
\description{
Short version of plotcli_box function.
}
