% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcli_wrappers.r
\name{pclil}
\alias{pclil}
\title{Short version of plotcli_line}
\usage{
pclil(
  y,
  x = NULL,
  plot_width = 80,
  plot_height = 40,
  x_label = "x",
  y_label = "y",
  color = NULL,
  braille = TRUE,
  name = "line",
  ...
)
}
\description{
Short version of plotcli_line function.
}
