% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{pixel_to_braille}
\alias{pixel_to_braille}
\title{Convert pixel coordinates to Braille cell and dot position}
\usage{
pixel_to_braille(px, py, canvas_rows, canvas_cols)
}
\arguments{
\item{px}{Pixel x coordinate (1-based)}

\item{py}{Pixel y coordinate (1-based, from top)}

\item{canvas_rows}{Number of character rows in canvas}

\item{canvas_cols}{Number of character columns in canvas}
}
\value{
List with cell_row, cell_col, dot_row, dot_col
}
\description{
Convert pixel coordinates to Braille cell and dot position
}
