% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{plotcli_options}
\alias{plotcli_options}
\title{Set global options for plotcli}
\usage{
plotcli_options(plot_width = 80, plot_height = 40, braille = TRUE)
}
\arguments{
\item{plot_width}{Default plot width (default: 80)}

\item{plot_height}{Default plot height (default: 40)}

\item{braille}{Default braille setting (default: TRUE)}
}
\description{
Set global options for plotcli
}
