% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.r
\name{+.plotcli}
\alias{+.plotcli}
\title{Overload the "+" operator for plotcli objects}
\usage{
\method{+}{plotcli}(plot1, plot2)
}
\arguments{
\item{plot1}{A plotcli object to be merged.}

\item{plot2}{A plotcli object to be merged.}
}
\value{
A new plotcli object containing the combined data from both objects.
}
\description{
This function overloads the "+" operator to merge two plotcli objects.
}
