% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.r
\name{rbind.plotcli}
\alias{rbind.plotcli}
\title{Combine two plotcli objects vertically}
\usage{
\method{rbind}{plotcli}(plot1, plot2)
}
\arguments{
\item{plot1}{A plotcli object to be combined.}

\item{plot2}{A plotcli object to be combined.}
}
\value{
A combined plot matrix.
}
\description{
This function combines two plotcli objects vertically, centering them horizontally.
}
