% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.r
\name{rbind_plots}
\alias{rbind_plots}
\title{Combine plot matrices vertically}
\usage{
rbind_plots(...)
}
\arguments{
\item{...}{A list of plot matrices to be combined.}
}
\value{
A combined plot matrix.
}
\description{
This function combines multiple plot matrices vertically, centering them horizontally.
}
