% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{remove_color_codes}
\alias{remove_color_codes}
\title{Remove color codes from a string}
\usage{
remove_color_codes(s)
}
\arguments{
\item{s}{The text string containing ANSI color codes.}
}
\value{
A text string with ANSI color codes removed.
}
\description{
This function removes ANSI color codes from a given text string.
}
\examples{
colored_text <- make_colored("Hello, world!", "blue")
remove_color_codes(colored_text)
}
