% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotcli_helpers.r
\name{render_faceted_plot}
\alias{render_faceted_plot}
\title{Render Faceted Plot}
\usage{
render_faceted_plot(built, facet_info, width, height, canvas_type, style_opts)
}
\arguments{
\item{built}{Result from ggplot_build}

\item{facet_info}{Facet information from get_facet_info}

\item{width}{Total width}

\item{height}{Total height}

\item{canvas_type}{Canvas type}

\item{style_opts}{Style options}
}
\description{
Render Faceted Plot
}
\keyword{internal}
