% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotcli_helpers.r
\name{render_single_panel}
\alias{render_single_panel}
\title{Render a single panel (non-faceted) plot}
\usage{
render_single_panel(built, width, height, canvas_type, style_opts)
}
\description{
Render a single panel (non-faceted) plot
}
\keyword{internal}
