% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_registry.r
\name{create_scales}
\alias{create_scales}
\title{Create Scale Object from ggplot_build data}
\usage{
create_scales(built, plot_width, plot_height, has_border = FALSE)
}
\arguments{
\item{built}{Result from ggplot_build()}

\item{plot_width}{Canvas pixel width}

\item{plot_height}{Canvas pixel height}

\item{has_border}{Whether a border will be drawn (adds padding)}
}
\value{
List with x_scale and y_scale functions
}
\description{
Create Scale Object from ggplot_build data
}
