% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcli_wrappers.r
\name{pclib}
\alias{pclib}
\title{Short version of plotcli_bar}
\usage{
pclib(
  y,
  x = NULL,
  plot_width = getOption("plotcli.plot_width", 80),
  plot_height = getOption("plotcli.plot_height", 40),
  x_label = "x",
  y_label = "y",
  color = NULL,
  braille = getOption("plotcli.braille", TRUE),
  name = "barplot",
  ...
)
}
\arguments{
\item{y}{A numeric vector of values}

\item{x}{A vector of categories}

\item{plot_width}{Width of the plot (default: 80)}

\item{plot_height}{Height of the plot (default: 40)}

\item{x_label}{Label for the x-axis (default: "x")}

\item{y_label}{Label for the y-axis (default: "y")}

\item{color}{Color of the plot elements (default: NULL)}

\item{braille}{Use Braille characters for the plot (default: TRUE)}

\item{name}{Name of the plot element (default: "barplot")}

\item{...}{Additional arguments passed to the plotcli$new() function}
}
\description{
Short version of plotcli_bar function.
}
\examples{
x <- 1:5
y <- c(10, 15, 8, 12, 6)
pclib(x, y)
}
