% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcli_wrappers.r
\name{pclil}
\alias{pclil}
\title{Short version of plotcli_line}
\usage{
pclil(
  y,
  x = NULL,
  plot_width = getOption("plotcli.plot_width", 80),
  plot_height = getOption("plotcli.plot_height", 40),
  x_label = "x",
  y_label = "y",
  color = NULL,
  braille = getOption("plotcli.braille", TRUE),
  name = "line",
  ...
)
}
\arguments{
\item{y}{A numeric vector of y values}

\item{x}{A numeric vector of x values}

\item{plot_width}{Width of the plot (default: 80)}

\item{plot_height}{Height of the plot (default: 40)}

\item{x_label}{Label for the x-axis (default: "x")}

\item{y_label}{Label for the y-axis (default: "y")}

\item{color}{Color of the plot elements (default: NULL)}

\item{braille}{Use Braille characters for the plot (default: TRUE)}

\item{name}{Name of the plot element (default: "line")}

\item{...}{Additional arguments passed to the plotcli$new() function}
}
\description{
Short version of plotcli_line function.
}
\examples{
x <- 1:10
y <- x^2
pclil(x, y)
}
