% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcli_wrappers.r
\name{pclis}
\alias{pclis}
\title{Short version of plotcli_scatter}
\usage{
pclis(
  y,
  x = NULL,
  plot_width = 80,
  plot_height = 40,
  x_label = "x",
  y_label = "y",
  color = NULL,
  braille = TRUE,
  name = "scatter",
  ...
)
}
\description{
Short version of plotcli_scatter function.
}
