% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{plotcli_options}
\alias{plotcli_options}
\title{Set global options for plotcli}
\usage{
plotcli_options(plot_width = 60, plot_height = 20, braille = FALSE)
}
\arguments{
\item{plot_width}{Default plot width (default: 60)}

\item{plot_height}{Default plot height (default: 20)}

\item{braille}{Default braille setting (default: FALSE)}
}
\description{
Set global options for plotcli
}
