% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.r
\name{rbind.plotcli}
\alias{rbind.plotcli}
\title{Generic function for combining plotcli objects vertically}
\usage{
\method{rbind}{plotcli}(..., deparse.level = 1)
}
\arguments{
\item{...}{plotcli objects to be combined.}

\item{deparse.level}{The deparsing level for the arguments.}
}
\value{
A combined plot matrix.
}
\description{
Generic function for combining plotcli objects vertically
}
